+++ CSK_Module_DeviceNetworkConfig/scripts/Configuration/DeviceNetworkConfig/helper/checkAPIs.lua
@@ -40,8 +40,15 @@ local function loadSpecificAPIs()
   Ethernet.Interface = require 'API.Ethernet.Interface'
 end
 
+-- Function to load DateTime APIs
+local function loadDateTimeAPIs()
+  -- If you want to check for specific APIs/functions supported on the device the module is running, place relevant APIs here
+  DateTime = require 'API.DateTime'
+end
+
 availableAPIs.default = xpcall(loadAPIs, debug.traceback) -- TRUE if all default APIs were loaded correctly
 availableAPIs.specific = xpcall(loadSpecificAPIs, debug.traceback) -- TRUE if all specific APIs were loaded correctly
+availableAPIs.dateTime = xpcall(loadDateTimeAPIs, debug.traceback) -- TRUE if DateTime API was loaded correctly
 
 return availableAPIs
 --**************************************************************************